Execute the reproduction test for more iterations
